xml/parsing/manakai/texts-2.dat

#data
  ab<p/>
#errors
1;1;no XML decl;;;s
1;3;text outside of root element;;;m
#document
| <{} p>

#data
|   
| ab<p/>
#errors
1;1;no XML decl;;;s
2;1;text outside of root element;;;m
#document
| <{} p>

#data
|   
|  ab<p/>
#errors
1;1;no XML decl;;;s
2;2;text outside of root element;;;m
#document
| <{} p>

#data
| 
| 
|  ab<p/>
#errors
1;1;no XML decl;;;s
3;2;text outside of root element;;;m
#document
| <{} p>
